代码疏通，接口优化

+++ src/main/java/com/tk/cratemanagement/dto/CrateContentDTO.java
@@ -1,13 +1,15 @@
 package com.tk.cratemanagement.dto;
 
 import com.tk.cratemanagement.domain.enumeration.CrateContentStatus;
+
+import java.math.BigDecimal;
 import java.time.Instant;
 
 public record CrateContentDTO(
     Long id,
     Long goodsId,
     String goodsName,
-    Double quantity,
+    BigDecimal quantity,
     CrateContentStatus status,
     Long locationId,
     String locationName,
